backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
groq==0.9.0
openai==1.30.1
google-generativeai==0.7.0
pydantic==2.7.1
httpx==0.27.0
